// hr-frontend/src/lib/api/generated/model/updateJobPositionRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UpdateJobPositionRequest {
  /**
     * @minLength 0
     * @maxLength 100
     * @pattern ^[A-Za-z0-9_-]+$
     */
  code: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  titleEn: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  titleAr: string;
  orgUnitId: number;
  gradeId?: number;
  /** @minimum 1 */
  positionLevel: number;
  reportsToPositionId?: number;
  /**
     * @minLength 0
     * @maxLength 5000
     */
  descriptionEn?: string;
  /**
     * @minLength 0
     * @maxLength 5000
     */
  descriptionAr?: string;
  open: boolean;
  active: boolean;
}